Refactoring folder changes
