+++ android_files/app/src/main/java/com/example/qradventure/ui/leaderboard/LeaderboardFragment.java
@@ -1,4 +1,4 @@
-package com.example.qradventure;
+package com.example.qradventure.ui.leaderboard;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.qradventure.R;
 
 /**
  * A simple {@link Fragment} subclass.
