+++ android_files/app/src/main/res/layout/fragment_leaderboard.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LeaderboardFragment">
+    tools:context=".ui.leaderboard.LeaderboardFragment">
 
     <TextView
         android:id="@+id/temp_leaderboard_text"
