+++ android_files/app/src/main/res/layout/fragment_scan.xml
@@ -5,7 +5,7 @@
     android:id="@+id/scan"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScanFragment">
+    tools:context=".ui.scan.ScanFragment">
 
     <TextView
         android:id="@+id/resultText"
